requirements.txt
openai
pdf2image
pillow
vllm
